Tests/reorganize: move email validation using custom validator test to own file
